New fix: replacing male & female (modloader) issues
Issues when replacing male and female dff/txd models through modloader still appeared.

By default: female actress loads as 1 and male actor loads as 2, but i was loading the male skin in the player as slot 4, as the void character in the same slot. So what i did was: if i need to change player skin as male, i load it as 2 (the same slot actor is using), if i need to change the player to female, i will load it in 1, and if i need to change player as void, i load it in another slot (in this time slot 10).

That fixed the issue for now... And i hope not encountering it again...

Looks like the securest way to work is not replacing skins through modloader, but keeping all of them in an .img, but game limits the amount of custom skins, unless we could use a limit adjuster. That task will be fot another day....